programming-guidelines: Add a page on available tooling
This is not meant as a tutorial on how to use any of the tools — merely
as an overview of the available tools for each task, and which ones are
currently recommended. It is expected that this page will evolve over
time as tools come and go.

Perhaps we should mention outdated tools for each task, and instruct the
user to avoid them?

https://bugzilla.gnome.org/show_bug.cgi?id=376123