platform-demos: Add a page on writing GSource implementations
This includes some example code which is discussed in the page. The code
comes with its own test suite which it is #included into, and the test
suite is run on `make check`.

The rules for compiling the test suite are entirely manual at the moment
to avoid introducing a dependency on a compiler to the gnome-devel-docs
module. They cannot be changed to use more conventional automake support
for building C programs, because the mere presence of a _SOURCES
variable causes automake to require a C compiler for dependency
generation.

https://bugzilla.gnome.org/show_bug.cgi?id=744717