programming-guidelines: long conditions are hard to read
The introduction of the page explains that code maintainability is
important, so having a note about a programming best practice is useful
I think.

The code example was more to explain the coding style, but with the
added note, we explain that we are aware that such long conditions are
not a good example of clean code.

https://bugzilla.gnome.org/show_bug.cgi?id=744521