programming-guidelines: add book references for writing good code
When possible, the link is the official web page for the book. As a
fallback, a link to wikipedia (if a page exists) or Amazon.

If only one book should be chosen for "writing good code" (the subject
of the page), it should be Code Complete IMHO. It covers a wide range of
subjects, from single lines of code to the general architecture of big
programs, with introductory chapters on other topics like refactoring,
debugging, performances, etc.

The Refactoring and Design Patterns books are also the references for
those subjects, which are important for code quality and
maintainability.

Finally, Object-Oriented Design Heuristics is also a really good book,
that contains many guidelines for writing OO code. It's IMHO a more
important book than Design Patterns, although the latter is more
well-known. The guidelines explained in the book apply to every OO code,
even for a small codebase. On the other hand some design patterns are
useful only for big codebases. Since every software must begin by a
small codebase, it's better to apply OO best-practices from the
start.

https://bugzilla.gnome.org/show_bug.cgi?id=747997