programming-guidelines: merging release commits
Absolutely requiring a linear Git history can be painful when doing
releases. Re-running make distcheck is not the funniest thing to do and
can take a long time (compiling the API documentation, running the
tests, etc). So I think a small exception to the rule is acceptable.

There was a discussion about that on the desktop-devel-list:
https://mail.gnome.org/archives/desktop-devel-list/2014-July/msg00027.html
the conclusion was that merging the release commit was the simplest
solution.

https://bugzilla.gnome.org/show_bug.cgi?id=757735